hablar.R: as_reliable_dte() edits to fix conversion of datetime to date
Fixes #1: For some reason, `strftime` removes a day for dates that have times at midnight. According to the documentation, changes have been made in R versions 4.2.0 and following:

> strftime is a wrapper for format.POSIXlt, and it and format.POSIXct first convert to class "POSIXlt" by calling [as.POSIXlt](http://127.0.0.1:30443/help/library/base/help/as.POSIXlt) (so they also work for class "[Date](http://127.0.0.1:30443/help/library/base/help/Date)"). Note that only that conversion depends on the time zone. Since R version 4.2.0, that as.POSIXlt() conversion now treats the non-finite numeric -Inf, Inf, NA and NaN differently (where previously all were treated as NA) and also the format() method for POSIXlt now treats these different non-finite times and dates analogously to type [double](http://127.0.0.1:30443/help/library/base/help/double).

Using `as.Date()` for variables belonging to the POSIXct class solves the problem, so the checking of the POSIXct class is not needed in `as_reliable_dte`.